+++ java/src/test/java/io/cucumber/prettyformatter/MessagesToSummaryWriterAcceptanceTest.java
@@ -2,7 +2,7 @@ package io.cucumber.prettyformatter;
 
 import io.cucumber.messages.NdjsonToMessageIterable;
 import io.cucumber.messages.types.Envelope;
-import io.cucumber.prettyformatter.MessagesToPrettyWriter.Builder;
+import io.cucumber.prettyformatter.MessagesToSummaryWriter.Builder;
 import org.junit.jupiter.api.Disabled;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.MethodSource;
@@ -18,37 +18,26 @@ import java.util.ArrayList;
 import java.util.LinkedHashMap;
 import java.util.List;
 import java.util.Map;
+import java.util.function.Consumer;
 import java.util.stream.Collectors;
 import java.util.stream.Stream;
 
 import static io.cucumber.prettyformatter.Jackson.OBJECT_MAPPER;
-import static io.cucumber.prettyformatter.MessagesToPrettyWriter.PrettyFeature.INCLUDE_ATTACHMENTS;
-import static io.cucumber.prettyformatter.MessagesToPrettyWriter.PrettyFeature.INCLUDE_FEATURE_LINE;
-import static io.cucumber.prettyformatter.MessagesToPrettyWriter.PrettyFeature.INCLUDE_RULE_LINE;
-import static io.cucumber.prettyformatter.MessagesToPrettyWriter.builder;
-import static io.cucumber.prettyformatter.TestTheme.demo;
+import static io.cucumber.prettyformatter.MessageOrderer.originalOrder;
+import static io.cucumber.prettyformatter.MessageOrderer.simulateParallelExecution;
+import static io.cucumber.prettyformatter.MessagesToSummaryWriter.builder;
 import static io.cucumber.prettyformatter.Theme.cucumber;
-import static io.cucumber.prettyformatter.Theme.none;
 import static io.cucumber.prettyformatter.Theme.plain;
 import static java.nio.file.Files.readAllBytes;
 import static org.assertj.core.api.Assertions.assertThat;
 
-class MessagesToPrettyWriterAcceptanceTest {
+class MessagesToSummaryWriterAcceptanceTest {
     private static final NdjsonToMessageIterable.Deserializer deserializer = (json) -> OBJECT_MAPPER.readValue(json, Envelope.class);
 
     static List<TestCase> acceptance() throws IOException {
-        Map<String, MessagesToPrettyWriter.Builder> themes = new LinkedHashMap<>();
+        Map<String, Builder> themes = new LinkedHashMap<>();
         themes.put("cucumber", builder().theme(cucumber()));
-        themes.put("demo", builder().theme(demo()));
         themes.put("plain", builder().theme(plain()));
-        themes.put("none", builder().theme(none()));
-        themes.put("exclude-features-and-rules", builder()
-                .theme(none())
-                .feature(INCLUDE_RULE_LINE, false)
-                .feature(INCLUDE_FEATURE_LINE, false));
-        themes.put("exclude-attachments", builder()
-                .theme(none())
-                .feature(INCLUDE_ATTACHMENTS, false));
 
         List<Path> sources = getSources();
 
@@ -68,14 +57,18 @@ class MessagesToPrettyWriterAcceptanceTest {
         }
     }
 
-    private static <T extends OutputStream> T writePrettyReport(TestCase testCase, T out, Builder builder) throws IOException {
+    private static <T extends OutputStream> T writeSummaryReport(TestCase testCase, T out, Builder builder, Consumer<List<Envelope>> orderer) throws IOException {
+        List<Envelope> messages = new ArrayList<>();
         try (InputStream in = Files.newInputStream(testCase.source)) {
             try (NdjsonToMessageIterable envelopes = new NdjsonToMessageIterable(in, deserializer)) {
-                try (MessagesToPrettyWriter writer = builder.build(out)) {
-                    for (Envelope envelope : envelopes) {
-                        writer.write(envelope);
-                    }
-                }
+                envelopes.forEach(messages::add);
+            }
+        }
+        orderer.accept(messages);
+
+        try (MessagesToSummaryWriter writer = builder.build(out)) {
+            for (Envelope envelope : messages) {
+                writer.write(envelope);
             }
         }
         return out;
@@ -84,16 +77,23 @@ class MessagesToPrettyWriterAcceptanceTest {
     @ParameterizedTest
     @MethodSource("acceptance")
     void test(TestCase testCase) throws IOException {
-        ByteArrayOutputStream bytes = writePrettyReport(testCase, new ByteArrayOutputStream(), testCase.builder);
+        ByteArrayOutputStream bytes = writeSummaryReport(testCase, new ByteArrayOutputStream(), testCase.builder, originalOrder());
         assertThat(bytes.toString()).isEqualToIgnoringNewLines(new String(readAllBytes(testCase.expected)));
     }
 
     @ParameterizedTest
     @MethodSource("acceptance")
+    void testWithSimulatedParallelExecution(TestCase testCase) throws IOException {
+        ByteArrayOutputStream bytes = writeSummaryReport(testCase, new ByteArrayOutputStream(), testCase.builder, simulateParallelExecution());
+        assertThat(bytes.toString()).isEqualToIgnoringNewLines(new String(readAllBytes(testCase.expected)));
+    }
+    
+    @ParameterizedTest
+    @MethodSource("acceptance")
     @Disabled
-    void updateExpected(TestCase testCase) throws IOException {
+    void updateExpectedFiles(TestCase testCase) throws IOException {
         try (OutputStream out = Files.newOutputStream(testCase.expected)) {
-            writePrettyReport(testCase, out, testCase.builder);
+            writeSummaryReport(testCase, out, testCase.builder, originalOrder());
             // Render output in console, easier to inspect results
             Files.copy(testCase.expected, System.out);
         }
@@ -113,7 +113,7 @@ class MessagesToPrettyWriterAcceptanceTest {
             this.builder = builder;
             String fileName = source.getFileName().toString();
             this.name = fileName.substring(0, fileName.lastIndexOf(".ndjson"));
-            this.expected = source.getParent().resolve(name + "." + themeName + ".pretty.log");
+            this.expected = source.getParent().resolve(name + "." + themeName + ".summary.log");
         }
 
         @Override