+++ CHANGELOG.md
@@ -6,6 +6,8 @@ The formatter is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- [Java] Message based summary printer ([#35](https://github.com/cucumber/pretty-formatter/pull/35))
 
 ## [2.2.0] - 2025-09-11
 ### Changed